+++ listopia-react/src/pages/SignIn.js
@@ -1,8 +1,12 @@
-import React from 'react';
+import React, { useEffect } from 'react';
 import styled from 'styled-components';
 import PropTypes from 'prop-types';
-import BounceLogo from './BounceLogo';
+import { connect } from 'react-redux';
+import BounceLogo from '../components/BounceLogo';
 import Theme from '../config/theme';
+
+import Categories from '../components/Categories';
+import { authenticate } from '../actions';
 
 const Container = styled.div`
   position: absolute;
@@ -70,4 +74,47 @@ Login.propTypes = {
   submitHandler: PropTypes.func.isRequired,
 };
 
-export default Login;
+const mapStateToProps = (state) => ({
+  authenticated: state.authenticated,
+});
+
+const mapDispatchToProps = (dispatch) => ({
+  authenticate: (userId) => dispatch(authenticate(userId)),
+});
+
+const IndexPage = ({ authenticated, authenticate }) => {
+  const submitHandler = (e) => {
+    e.preventDefault();
+    authenticate(e.target.name.value);
+    localStorage.setItem('userId', e.target.name.value);
+    document.activeElement.blur();
+  };
+
+  const checkSession = () => {
+    const userId = localStorage.getItem('userId');
+    authenticate(userId);
+  };
+
+  useEffect(() => {
+    if (authenticated === 'pending') {
+      new Promise((resolve) => {
+        setTimeout(() => {
+          resolve('resolved');
+        }, 1500);
+      }).then(() => checkSession());
+    }
+  });
+
+  return (
+    <div>
+      <Login authenticated={authenticated} submitHandler={submitHandler} />
+    </div>
+  );
+};
+
+IndexPage.propTypes = {
+  authenticated: PropTypes.string.isRequired,
+  authenticate: PropTypes.func.isRequired,
+};
+
+export default connect(mapStateToProps, mapDispatchToProps)(IndexPage);
